% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariates.R
\docType{data}
\name{covariates}
\alias{covariates}
\title{Example dataframe}
\format{
A data frame with 3720 rows and 13 variables: {
variable names:
country_name
ISO
region
year
GDP1000
OECD
OECD2023
popdens
urban
agedep
male_edu
temp
rainfall1000

}
}
\usage{
covariates
}
\description{
Sample data with multiple variables with names of varying lengths, spacing,
and captilization. This dataframe also contains variables with similar names
which the function deals with by adding col# to the end of repeated named
variables
}
\keyword{datasets}
