% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Example simulated dataframe to test codedtitles package}
\format{
A simulated data frame with 20 rows and 36 variables: {
Example variables names are included such as:
   "?@%#%" <- a value with all special characters,
   "Attack_Rate" and "Attack_Rate_Total" which share similar stems
   "Mort$Total%_" is a variable with a mix of meaningful word stems
     and special characters

}
}
\usage{
df
}
\description{
Sample data with multiple variables with names of varying lengths, spacing,
and capitalization. This dataframe also contains variables with similar names
which the function deals with by adding col# to the end of repeated named
variables
}
\keyword{datasets}
