% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codedtitle.R
\name{codevar}
\alias{codevar}
\title{Code Variable Names of Dataframe}
\usage{
codevar(data, strlength)
}
\arguments{
\item{dataframe}{with named columns}
}
\value{
dataframe with recoded column names
}
\description{
Take a dataset and change column names to manageable
coded variables names based on a desired string-length. This function also
removes all spaces and makes all characters lower-case.
}
\examples{
codetitle(data, strlength = 3)
}
\author{
Mohsyn Imran Malik
}
